Finished Dev Version 1.0. Ready to real tests.
- Added view to create and delete categories.
- Update README
- Added requirements.txt file
- Added authentication check in some views.

- Added authentication check in some point

+++ Materials/views.py
@@ -7,10 +7,11 @@ Responsible for views related to the models defined in Materials/models.py.
 from django.shortcuts import render, HttpResponse, redirect
 from .models import Apostila, Project, Categoria
 from django.urls import reverse_lazy
+from django.http import HttpResponseForbidden
 
 from ArtHeart.settings import MEDIA_ROOT
 import os
-from .forms import ApostilaUpload, ProjectUpload
+from .forms import ApostilaUpload, ProjectUpload, CategoryUpload
 
 # Create your views here.
 def search(request):
@@ -49,46 +50,50 @@ def detail(request, pk):
 
     Template for this view is 'Materials/detail.html'
     """
-    pk = int(pk)
-    if request.method == 'GET':
-        if pk != -1:
-            ap = Apostila.objects.get(pk=pk)
-            form = ApostilaUpload(instance=ap)
-            data = {'doc':ap, 'form':form, 'edit':True }
-        else:
-            form = ApostilaUpload()
-            data = { 'form':form, 'edit':False }
-        return render(request, 'Materials/detail.html', data)
-    else:
-        if pk != -1:
-            if 'file-clear' in request.POST or 'file' in request.FILES:
-                deleteFile(pk)
-            form = ApostilaUpload(request.POST or None, request.FILES or None, instance=Apostila.objects.get(pk=pk))
-        else:
-            form = ApostilaUpload(request.POST or None, request.FILES or None)
-        if form.is_valid() :
-            entry = form.save()
-            cats = request.POST.getlist('categories')
-            selected = []
-            for c in cats:
-                selected.append(Categoria.objects.get(pk=c))
-
-            entry.categories.set(selected)
-            entry.save()
-            return redirect(reverse_lazy('url_search'))
-        else:
-            data = {'form':form}
+    if request.user.is_authenticated:
+        pk = int(pk)
+        if request.method == 'GET':
             if pk != -1:
-                data['doc'] = Apostila.objects.get(pk=pk)
-                data['edit'] = True
+                ap = Apostila.objects.get(pk=pk)
+                form = ApostilaUpload(instance=ap)
+                data = {'doc':ap, 'form':form, 'edit':True }
             else:
-                data['edit'] = False
-            return render(request, 'Materials/details.html', )
+                form = ApostilaUpload()
+                data = { 'form':form, 'edit':False }
+            return render(request, 'Materials/detail.html', data)
+        else:
+            if pk != -1:
+                if 'file-clear' in request.POST or 'file' in request.FILES:
+                    deleteFile(pk)
+                form = ApostilaUpload(request.POST or None, request.FILES or None, instance=Apostila.objects.get(pk=pk))
+            else:
+                form = ApostilaUpload(request.POST or None, request.FILES or None)
+            if form.is_valid() :
+                entry = form.save()
+                cats = request.POST.getlist('categories')
+                selected = []
+                for c in cats:
+                    selected.append(Categoria.objects.get(pk=c))
+
+                entry.categories.set(selected)
+                entry.save()
+                return redirect(reverse_lazy('url_search'))
+            else:
+                data = {'form':form}
+                if pk != -1:
+                    data['doc'] = Apostila.objects.get(pk=pk)
+                    data['edit'] = True
+                else:
+                    data['edit'] = False
+                return render(request, 'Materials/details.html', )
+    else:
+        return HttpResponseForbidden()
 
 def projects(request):
     """
     Shows available Projects. Also possible to add new Project.
     Quick glance at Project's title, client, and state of completion.
+    User that are not authenticated CANNOT create new projects
 
     Template for this view is 'Materials/projects.html'
     Links to edit_project view.
@@ -96,7 +101,7 @@ def projects(request):
     data = {}
     data['projects'] = Project.objects.all()
     data['form'] = ProjectUpload()
-    if request.method == 'POST':
+    if request.method == 'POST' and request.user.is_authenticated:
         form = ProjectUpload(request.POST or None)
         if form.is_valid():
             form.save()
@@ -107,26 +112,59 @@ def projects(request):
         return render(request, 'Materials/projects.html', data)
 
 def edit_project(request, pk):
-    p = Project.objects.get(pk=pk)
-    form = ProjectUpload(instance=p)
-    entries = p.apostila_set.all()
-    data = {'doc':p, 'form':form, 'entries':entries}
-    if request.method == 'POST':
-        form = ProjectUpload(request.POST, instance=p)
-        if form.is_valid():
-            entry = form.save(commit=False)
-            if 'completed' in request.POST:
-                entry.completed = True
-            entry.save()
-            return redirect(reverse_lazy('url_projects'))
-        else:
-            data['form'] = form
-            return render(request, 'Materials/edit_project.html', data)
-    return render(request, 'Materials/edit_project.html', data)
+    """
+    Manages de editing of an existing Project.
+    On successful operation returns to projects view.
+
+    Template for this view is 'Materials/edit_project.html'
+    """
+    if request.user.is_authenticated:
+        p = Project.objects.get(pk=pk)
+        form = ProjectUpload(instance=p)
+        entries = p.apostila_set.all()
+        data = {'doc':p, 'form':form, 'entries':entries}
+        if request.method == 'POST':
+            form = ProjectUpload(request.POST, instance=p)
+            if form.is_valid():
+                entry = form.save(commit=False)
+                if 'completed' in request.POST:
+                    entry.completed = True
+                entry.save()
+                return redirect(reverse_lazy('url_projects'))
+            else:
+                data['form'] = form
+                return render(request, 'Materials/edit_project.html', data)
+        return render(request, 'Materials/edit_project.html', data)
+    else:
+        return HttpResponseForbidden()
         
+def category(request, pk):
+    """
+    Manages creation and Deletion of Categoria.
+
+    Template for this view is 'Materials/category.html'
+    """
+    if request.method == 'GET':
+        if pk != None:
+            try:
+                Categoria.objects.get(pk=pk).delete()
+            except:
+                pass
+    else:
+        form = CategoryUpload(request.POST or None)
+        if form.is_valid():
+            form.save()
+
+    cats = Categoria.objects.all()
+    form = CategoryUpload()
+    data = {'cat':cats, 'form':form}
+    return render(request, 'Materials/category.html', data)
+
 def deleteFile(pk):
     """
     Deletes a file referenced by an Apostila.
+
+    Has no template or URL.
     """
     ap = Apostila.objects.get(pk=pk)
     try:
@@ -137,6 +175,8 @@ def deleteFile(pk):
 def delete(request, pk):
     """
     Deleta uma Apostila do banco de dados e qualquer arquivo referênciado por ele.
+
+    Has no template.
     """
     if request.user.is_authenticated:
         doc = Apostila.objects.get(pk=pk)
@@ -145,6 +185,12 @@ def delete(request, pk):
     return redirect('url_search')
 
 def delete_project(request, pk):
+    """
+    Deletes a Project along with ALL the Apostilas related to it.
+    If the Apostilas have files they are deleted as well.
+
+    Has no template.
+    """
     if request.user.is_authenticated:
         p = Project.objects.get(pk=pk)
         aps = p.apostila_set.all()
